Add event ping to EventsToAmplitude and devtools config
