+++ src/main/resources/schemas/devtoolsConfigFile.json
@@ -1,7 +1,8 @@
 {
   "filters": {
     "docType": [
-      "main"
+      "main",
+      "event"
     ],
     "appName": [
       "Firefox"
